requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
colorzero==1.1
cryptography==37.0.2
# discord==1.7.3
discord==2.3.2
# discord.py==1.7.3
discord.py==2.3.2
distro==1.5.0
ffmpeg==1.4
idna==2.10
jsonschema==3.2.0
mongoengine==0.24.1
multidict==6.0.2
paramiko==2.11.0
praw==7.7.0
prawcore==2.3.0
pymongo==4.1.1
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.20.0
PyYAML==6.0.1
requests==2.25.1
six==1.16.0
ssh-import-id==5.10
typing-extensions==4.2.0
update-checker==0.18.0
urllib3==1.26.5
websocket-client==1.3.2
yamllint==1.26.0
youtube-dl==2021.12.17
